+++ src/misc.ts
@@ -1,3 +1,1 @@
-export const NEW_LINE = "#-##--&--##-#";
-
 export type NewLineType = '\r\n'|'\n';